+++ src/api/sk_const.h
@@ -30,6 +30,12 @@ struct sk_const_t;
 #define SK_USER_PREFIX2                      "/usr/lib"
 #define SK_USER_LIBNAME_MAX                  1024
 
+// network
+//  workflow trigger
+#define SK_MAX_LISTEN_BACKLOG                (1024)
+#define SK_DEFAULT_READBUF_LEN               (65535)
+#define SK_DEFAULT_READBUF_FACTOR            (2)
+
 // scheduler
 #define SK_SCHED_PULL_NUM                    65535
 #define SK_SCHED_MAX_IO_BRIDGE               48
@@ -47,7 +53,7 @@ struct sk_const_t;
 
 #define SK_CONFIG_NO_PORT                    (-1)
 #define SK_CONFIG_DEFAULT_CMD_PORT           (7759)
-#define SK_CONFIG_VALUE_MAXLEN               256
+#define SK_CONFIG_VALUE_MAXLEN               (256)
 
 // ep pool
 #define SK_EP_POOL_MAX                       (1024)