+++ src/api/sk_entity_mgr.h
@@ -25,9 +25,11 @@ typedef struct sk_entity_mgr_stat_t {
     int inactive;
 
     int entity_none;
-    int entity_net;
+    int entity_sock_v4tcp;  // Client entity, v4 tcp
+    int entity_sock_v4udp;  // Client entity, v4 udp
     int entity_timer;
-    int entity_ep;
+    int entity_ep_v4tcp;
+    int entity_ep_v4udp;
     int entity_ep_timer;
     int entity_ep_txn_timer;
 } sk_entity_mgr_stat_t;