+++ src/api/sk_trigger_utils.h
@@ -1,0 +1,9 @@
+#ifndef SK_TRIGGER_UTILS_H
+#define SK_TRIGGER_UTILS_H
+
+#include "api/sk_entity.h"
+
+ssize_t sk_trigger_util_unpack(sk_entity_t* entity);
+
+#endif
+