+++ src/components/sk_service.c
@@ -714,7 +714,7 @@ sk_service_job_create(sk_service_t*       service,
     // 1. Create timer callback data
     srv_jobdata_t* jobdata = calloc(1, sizeof(*jobdata));
     jobdata->service = service;
-    jobdata->entity  = sk_entity_create(NULL, SK_ENTITY_TAG_TIMER);
+    jobdata->entity  = sk_entity_create(NULL, SK_ENTITY_TIMER);
     jobdata->job     = job;
     jobdata->ud      = ud;
     jobdata->bidx    = bidx;