+++ src/pto/sk_pto_stdin_start.c
@@ -9,8 +9,9 @@
 #include "api/sk_pto.h"
 #include "api/sk_log.h"
 #include "api/sk_entity.h"
-#include "api/sk_entity_util.h"
 #include "api/sk_sched.h"
+#include "api/sk_entity_util.h"
+#include "api/sk_trigger_utils.h"
 
 static
 void _read_cb(fev_state* fev, fev_buff* evbuff, void* arg)
@@ -26,7 +27,7 @@ void _read_cb(fev_state* fev, fev_buff* evbuff, void* arg)
         return;
     }
 
-    sk_entity_util_unpack(fev, evbuff, entity);
+    sk_trigger_util_unpack(entity);
 }
 
 static